pathomAI-main/backend/requirements.txt
fastapi>=0.113.0
uvicorn[standard]>=0.30.0
openai>=1.60.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
python-multipart>=0.0.9
ffmpeg-python>=0.2.0
httpx>=0.27.2
python-dotenv>=1.0.1
email-validator>=2.2.0
